backend/requirements.txt
fastapi	
uvicorn	
faiss-cpu
openai	
pydantic
speechrecognition	
pyttsx3	
python-dotenv	
streamlit	
requests	